requirements.txt
# Core dependencies
chainlit
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# ADK and Google AI (for future integration)
# google-generativeai==0.3.0
# google-cloud-aiplatform==1.38.0

# Database
google-cloud-firestore>=2.14.0  # For production persistent storage

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6
